agent-config persist/populate, renew cert flow
The renew cert flow was a little bit more complicated than it needed to
be. Now we only use a timer to trigger the renewals, this is simpler but
will lead to more "unnecessary" renewals.

The agent now also saves and loads its config to/from disk so it will be
persistent.

Some general fixes like logging unhandled errors here and there