outtune: extract public key from browser db
This fixes bug with multiple browsers, as we previously only kept track
of one public key.
Also removes the old public key from disk at it is no longer used.

Also checks if we have an existing cert with nickname naisdevice instead
of checking if we have a key, as it turns out orphaned keys are a normal
thing in nss databases - and very hard to recover from without wiping
the entire keystore.