Make homebrew and scoop workflows succeed even if there are no new files to commit
We run the workflows on all builds, but only a few builds results in actual new releases.
Previously, the workflow would add a new file, commit and push it, and if there was no new file, the workflow fails.
With this change, the workflow will only attempt to commit and push if there is a new file to add, otherwise do nothing.